+++ src/pages/Researcher.jsx
@@ -1,5 +1,5 @@
 import DashboardLayout from '../layouts/DashboardLayout'
-import UserTable from '../components/common/UserTable/UserTable'
+import ResearcherTable from '../components/researcher/ResearcherTable'
 
 const Researcher = () => {
   return (
@@ -9,7 +9,7 @@ const Researcher = () => {
           <h1 className="text-2xl font-semibold text-gray-900">Peneliti</h1>
         </div>
         <div className="max-w-7xl mx-auto px-4 sm:px-6 md:px-8">
-          <UserTable />
+          <ResearcherTable />
         </div>
       </div>
     </DashboardLayout>
